fix: update arn builder method and its usage
